Move index page to form page
